+++ src/main/java/com/base_spring_boot/com/tmoto/base/service/Service.java
@@ -7,5 +7,5 @@ import org.springframework.util.MultiValueMap;
 public interface Service<T> {
     Page<T> find(Pageable page, MultiValueMap<String, String> criteria);
 
-    String getUserAllowedMethodHeaders(Long id);
+    String getUserAllowedMethodHeaders(Integer id);
 }
